requirements.txt
flet
pandas
librosa 
hmmlearn
openai-whisper
requests
ffmpeg-python
scipy
